pyproject.toml
[project]
name = "adk-ollama-CKM"
version = "0.1.0"
description = "Google Agent ADK CKM using Ollama ministral-3:14b"
requires-python = ">=3.10"
dependencies = [
    "google-adk[all]>=1.0.0",
    "litellm>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

